+++ client/src/main/java/il/cshaifasweng/OCSFMediatorExample/client/RegisteredCustomerController.java
@@ -255,7 +255,7 @@ public class RegisteredCustomerController {
         orderExit.setCellValueFactory(new PropertyValueFactory<>("exiting"));
         orderLicense.setCellValueFactory(new PropertyValueFactory<>("plateNum"));
         orderParchaseDate.setCellValueFactory(new PropertyValueFactory<>("date"));
-        orderPLotID.setCellValueFactory(data -> new SimpleObjectProperty<>(Integer.toString(data.getValue().getParkingLotID().getId())));
+        orderPLotID.setCellValueFactory(data -> new SimpleObjectProperty<>(Long.toString(data.getValue().getParkingLotID().getId())));
         orderPricePaid.setCellValueFactory(data -> new SimpleObjectProperty<>(Double.toString(data.getValue().getValue())));
         observableOrders.forEach(ordersTable.getItems()::add);
     }
